move updated_at check out of github_metadata module, to awesome_lint module - fixes https://github.com/nodiscc/hecat/issues/47
