+++ hecat/processors/__init__.py
@@ -1,4 +1,4 @@
 """processors"""
-from .github_metadata import add_github_metadata, check_github_last_updated
-from .awesome_lint import awesome_lint
+from .github_metadata import add_github_metadata
+from .awesome_lint import awesome_lint, check_github_last_updated
 from .download_media import download_media
